fixup: move stuff to testing file
